Move: useMaxDisplayCount 훅 BadgeCounter 폴더 안으로 위치 변경
