+++ components/Navbar/BadgeCounter/index.tsx
@@ -5,7 +5,7 @@ my dashboard에서 member들 보여주는 카운터
 import { FC } from "react";
 import UserBadge from "@/components/Navbar/UserBadge/UserBadge";
 import { mockMembersData } from "../MockData";
-import useMaxDisplayCount from "@/hooks/useMaxDisplayCount";
+import useMaxDisplayCount from "./useMaxDisplayCount";
 
 const BadgeCounter: FC = () => {
   const { members, totalCount } = mockMembersData;
