+++ components/Navbar/BadgeCounter/useMaxDisplayCount.ts
